+++ serverside/roboto/src/main/kotlin/AndroidAttestationConfiguration.kt
@@ -211,6 +211,12 @@ data class AndroidAttestationConfiguration @JvmOverloads constructor(
     val enableSoftwareAttestation: Boolean = false,
 
     /**
+     * [Mandates Remote Key Provisioning (RKP)](https://source.android.com/docs/core/ota/modular-system/remote-key-provisioning)
+     * for attestation checks to pass
+     */
+    val requireRemoteKeyProvisioning: Boolean = false,
+
+    /**
      * HTTP Proxy URL formatted as `http(s)://proxy-domain:port`
      */
     val httpProxy: String? = null,
@@ -316,6 +322,12 @@ data class AndroidAttestationConfiguration @JvmOverloads constructor(
          * HTTP Proxy URL formatted as `http(s)://proxy-domain:port`
          */
         httpProxy: String? = null,
+
+        /**
+         * [Mandates Remote Key Provisioning (RKP)](https://source.android.com/docs/core/ota/modular-system/remote-key-provisioning)
+         * for attestation checks to pass
+         */
+        requireRemoteKeyProvisioning: Boolean = false,
     ) : this(
         listOf(singleApp),
         androidVersion = androidVersion,
@@ -331,7 +343,8 @@ data class AndroidAttestationConfiguration @JvmOverloads constructor(
         disableHardwareAttestation = disableHardwareAttestation,
         enableNougatAttestation = enableNougatAttestation,
         enableSoftwareAttestation = enableSoftwareAttestation,
-        httpProxy = httpProxy
+        httpProxy = httpProxy,
+        requireRemoteKeyProvisioning = requireRemoteKeyProvisioning,
     )
 
     /**
@@ -440,6 +453,12 @@ data class AndroidAttestationConfiguration @JvmOverloads constructor(
          * HTTP Proxy URL formatted as `http(s)://proxy-domain:port`
          */
         httpProxy: String? = null,
+
+        /**
+         * [Mandates Remote Key Provisioning (RKP)](https://source.android.com/docs/core/ota/modular-system/remote-key-provisioning)
+         * for attestation checks to pass
+         */
+        requireRemoteKeyProvisioning: Boolean = false,
     ) : this(
         applications = apps,
         androidVersion = version,
@@ -456,6 +475,7 @@ data class AndroidAttestationConfiguration @JvmOverloads constructor(
         enableNougatAttestation = enableNougatAttestation,
         enableSoftwareAttestation = enableSoftwareAttestation,
         httpProxy = httpProxy,
+        requireRemoteKeyProvisioning = requireRemoteKeyProvisioning,
     )
 
     /**
@@ -506,6 +526,12 @@ data class AndroidAttestationConfiguration @JvmOverloads constructor(
          */
         val trustAnchorOverrides: Set<@Serializable(with = PubKeyBasePemSerializer::class) PublicKey>? = null,
 
+        /**
+         * [Mandates Remote Key Provisioning (RKP)](https://source.android.com/docs/core/ota/modular-system/remote-key-provisioning)
+         * for attestation checks to pass
+         */
+        val requireRemoteKeyProvisioningOverride: Boolean? = null,
+
         ) {
         init {
             if (signatureDigests.isEmpty()) throw object :
@@ -544,6 +570,8 @@ data class AndroidAttestationConfiguration @JvmOverloads constructor(
 
             private var trustAnchorOverrides: Set<PublicKey>? = null
 
+            var requireRemoteKeyProvisioningOverride: Boolean? = null
+
             /**
              * @see AppData.appVersion
              */
@@ -564,6 +592,13 @@ data class AndroidAttestationConfiguration @JvmOverloads constructor(
              */
             fun overrideTrustAnchors(trustAnchors: Set<PublicKey>) = apply { trustAnchorOverrides = trustAnchors }
 
+            /**
+             * [Mandates Remote Key Provisioning (RKP)](https://source.android.com/docs/core/ota/modular-system/remote-key-provisioning)
+             * for attestation checks to pass
+             */
+            fun overrideRequireRemoteProvisioning(required: Boolean) =
+                apply { requireRemoteKeyProvisioningOverride = required }
+
             fun build() =
                 AppData(
                     packageName,
@@ -571,7 +606,8 @@ data class AndroidAttestationConfiguration @JvmOverloads constructor(
                     appVersion,
                     androidVersionOverride,
                     patchLevelOverride,
-                    trustAnchorOverrides
+                    trustAnchorOverrides,
+                    requireRemoteKeyProvisioningOverride,
                 )
         }
 
@@ -605,6 +641,8 @@ data class AndroidAttestationConfiguration @JvmOverloads constructor(
 
             if (patchLevelOverride != other.patchLevelOverride) return false
 
+            if(requireRemoteKeyProvisioningOverride != other.requireRemoteKeyProvisioningOverride) return false
+
             return true
         }
 
@@ -615,6 +653,8 @@ data class AndroidAttestationConfiguration @JvmOverloads constructor(
             result = 31 * result + packageName.hashCode()
             result = 31 * result + signatureDigests.hashCode()
             result = 31 * result + (patchLevelOverride?.hashCode() ?: 0)
+            result = 31 * result + trustAnchorOverrides.hashCode()
+            result = 31 * result + requireRemoteKeyProvisioningOverride.hashCode()
             return result
         }
 
@@ -656,6 +696,7 @@ data class AndroidAttestationConfiguration @JvmOverloads constructor(
         private var enableSwAttestation: Boolean = false
         private var enableNougatAttestation: Boolean = false
         private var httpProxy: String? = null
+        private var requireRemoteKeyProvisioning: Boolean = false
 
         /**
          * specifies a minimum Android version
@@ -746,6 +787,11 @@ data class AndroidAttestationConfiguration @JvmOverloads constructor(
          */
         fun httpProxy(url: String) = apply { httpProxy = url }
 
+        /**
+         * @see at.asitplus.attestation.android.AndroidAttestationConfiguration.requireRemoteKeyProvisioning
+         */
+        fun requireRemoteKeyProvisioning(required: Boolean) = apply { requireRemoteKeyProvisioning = required }
+
         fun build() = AndroidAttestationConfiguration(
             applications = applications,
             androidVersion = androidVersion,
@@ -762,6 +808,7 @@ data class AndroidAttestationConfiguration @JvmOverloads constructor(
             enableSoftwareAttestation = enableSwAttestation,
             enableNougatAttestation = enableNougatAttestation,
             httpProxy = httpProxy,
+            requireRemoteKeyProvisioning = requireRemoteKeyProvisioning,
         )
 
     }
@@ -813,6 +860,8 @@ data class AndroidAttestationConfiguration @JvmOverloads constructor(
 
         if (httpProxy != other.httpProxy) return false
 
+        if (requireRemoteKeyProvisioning != other.requireRemoteKeyProvisioning) return false
+
         return true
     }
 
@@ -833,6 +882,7 @@ data class AndroidAttestationConfiguration @JvmOverloads constructor(
         result = 31 * result + hardwareAttestationTrustAnchors.hashCode()
         result = 31 * result + softwareAttestationTrustAnchors.hashCode()
         result = 31 * result + (httpProxy?.hashCode() ?: 0)
+        result = 31 * result + requireRemoteKeyProvisioning.hashCode()
         return result
     }
 }